+++ app/src/main/java/org/gdg/frisbee/android/onboarding/FirstStartStep2Fragment.java
@@ -1,0 +1,248 @@
+/*
+ * Copyright 2013-2015 The GDG Frisbee Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gdg.frisbee.android.onboarding;
+
+import android.app.Activity;
+import android.app.PendingIntent;
+import android.content.Intent;
+import android.content.IntentSender;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.games.Games;
+import com.google.android.gms.plus.Plus;
+
+import org.gdg.frisbee.android.R;
+import org.gdg.frisbee.android.fragment.BaseFragment;
+
+import butterknife.ButterKnife;
+import butterknife.InjectView;
+
+public class FirstStartStep2Fragment extends BaseFragment implements
+        GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
+
+    private static final int STATE_DEFAULT = 0;
+    private static final int STATE_SIGN_IN = 1;
+    private static final int STATE_IN_PROGRESS = 2;
+
+    private static final int RC_SIGN_IN = 0;
+
+    private static final int DIALOG_PLAY_SERVICES_ERROR = 0;
+
+    private static final String SAVED_PROGRESS = "sign_in_progress";
+    @InjectView(R.id.googleSignin)
+    Button mSignInButton;
+    @InjectView(R.id.skipSignin)
+    Button mSkipSignin;
+    private GoogleApiClient mGoogleApiClient;
+    // We use mSignInProgress to track whether user has clicked sign in.
+    // mSignInProgress can be one of three values:
+    //
+    //       STATE_DEFAULT: The default state of the application before the user
+    //                      has clicked 'sign in', or after they have clicked
+    //                      'sign out'.  In this state we will not attempt to
+    //                      resolve sign in errors and so will display our
+    //                      Activity in a signed out state.
+    //       STATE_SIGN_IN: This state indicates that the user has clicked 'sign
+    //                      in', so resolve successive errors preventing sign in
+    //                      until the user has successfully authorized an account
+    //                      for our app.
+    //   STATE_IN_PROGRESS: This state indicates that we have started an intent to
+    //                      resolve an error, and so we should not start further
+    //                      intents until the current intent completes.
+    private int mSignInProgress;
+    // Used to store the PendingIntent most recently returned by Google Play
+    // services until the user clicks 'sign in'.
+    private PendingIntent mSignInIntent;
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        if (savedInstanceState != null) {
+            mSignInProgress = savedInstanceState
+                    .getInt(SAVED_PROGRESS, STATE_DEFAULT);
+        }
+    }
+
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+
+        mGoogleApiClient = new GoogleApiClient.Builder(getActivity())
+                .addConnectionCallbacks(this)
+                .addOnConnectionFailedListener(this)
+                .addApi(Plus.API)
+                .addScope(Plus.SCOPE_PLUS_LOGIN)
+                .addScope(Plus.SCOPE_PLUS_PROFILE)
+                .addScope(Games.SCOPE_GAMES)
+                .build();
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        mGoogleApiClient.connect();
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View v = inflater.inflate(R.layout.fragment_welcome_step2, container, false);
+        ButterKnife.inject(this, v);
+        return v;
+    }
+
+    @Override
+    public void onViewCreated(View view, Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        mSignInButton.setOnClickListener(new View.OnClickListener() {
+
+            @Override
+            public void onClick(View view) {
+                if (getActivity() instanceof Step2Listener) {
+                    resolveSignInError();
+                }
+            }
+        });
+
+        mSkipSignin.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                if (getActivity() instanceof Step2Listener) {
+                    ((Step2Listener) getActivity()).onSkippedSignIn();
+                }
+            }
+        });
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+
+        if (mGoogleApiClient.isConnected()) {
+            mGoogleApiClient.disconnect();
+        }
+    }
+
+    @Override
+    public void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+        outState.putInt(SAVED_PROGRESS, mSignInProgress);
+    }
+
+    @Override
+    public void onConnected(Bundle bundle) {
+        if (getActivity() instanceof Step2Listener) {
+            ((Step2Listener) getActivity()).onSignedIn(mGoogleApiClient, Plus.AccountApi.getAccountName(mGoogleApiClient));
+        }
+    }
+
+    @Override
+    public void onConnectionSuspended(int i) {
+        mGoogleApiClient.connect();
+    }
+
+    @Override
+    public void onConnectionFailed(ConnectionResult result) {
+        if (mSignInProgress != STATE_IN_PROGRESS) {
+            // We do not have an intent in progress so we should store the latest
+            // error resolution intent for use when the sign in button is clicked.
+            mSignInIntent = result.getResolution();
+
+            if (mSignInProgress == STATE_SIGN_IN) {
+                // STATE_SIGN_IN indicates the user already clicked the sign in button
+                // so we should continue processing errors until the user is signed in
+                // or they click cancel.
+                resolveSignInError();
+            }
+        }
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        switch (requestCode) {
+            case RC_SIGN_IN:
+                if (resultCode == Activity.RESULT_OK) {
+                    // If the error resolution was successful we should continue
+                    // processing errors.
+                    mSignInProgress = STATE_SIGN_IN;
+                } else {
+                    // If the error resolution was not successful or the user canceled,
+                    // we should stop processing errors.
+                    mSignInProgress = STATE_DEFAULT;
+                }
+
+                if (mGoogleApiClient != null && !mGoogleApiClient.isConnecting()) {
+                    // If Google Play services resolved the issue with a dialog then
+                    // onStart is not called so we need to re-attempt connection here.
+                    mGoogleApiClient.connect();
+                }
+                break;
+        }
+    }
+
+    /* Starts an appropriate intent or dialog for user interaction to resolve
+       * the current error preventing the user from being signed in.  This could
+       * be a dialog allowing the user to select an account, an activity allowing
+       * the user to consent to the permissions being requested by your app, a
+       * setting to enable device networking, etc.
+       */
+    private void resolveSignInError() {
+        if (mSignInIntent != null) {
+            // We have an intent which will allow our user to sign in or
+            // resolve an error.  For example if the user needs to
+            // select an account to sign in with, or if they need to consent
+            // to the permissions your app is requesting.
+
+            try {
+                // Send the pending intent that we stored on the most recent
+                // OnConnectionFailed callback.  This will allow the user to
+                // resolve the error currently preventing our connection to
+                // Google Play services.
+                mSignInProgress = STATE_IN_PROGRESS;
+                getActivity().startIntentSenderForResult(mSignInIntent.getIntentSender(),
+                        RC_SIGN_IN, null, 0, 0, 0);
+            } catch (IntentSender.SendIntentException e) {
+                // The intent was canceled before it was sent.  Attempt to connect to
+                // get an updated ConnectionResult.
+                mSignInProgress = STATE_SIGN_IN;
+                mGoogleApiClient.connect();
+            }
+        } else {
+            // Google Play services wasn't able to provide an intent for some
+            // error types, so we show the default Google Play services error
+            // dialog which may still start an intent on our behalf if the
+            // user can resolve the issue.
+            //TODO: replace this with the right Dialog implementation.
+            getActivity().showDialog(DIALOG_PLAY_SERVICES_ERROR);
+        }
+    }
+
+    public interface Step2Listener {
+        void onSignedIn(GoogleApiClient client, String accountName);
+
+        void onSkippedSignIn();
+    }
+}
